requirements.txt
python-telegram-bot==20.7
python-dotenv==1.0.0
pytesseract==0.3.10
pdf2image==1.17.0
pillow==10.2.0
opencv-python-headless==4.12.0.88
numpy==1.26.4
# pillow-heif==0.15.0  # HEIC webde de olsun istiyorsan aç
